refactor: move director out of main.py
Move all the pre-setup logic to its own module and define a director class to manage it.
